--- client/src/services/event-service.js
@@ -1,6 +1,0 @@
-import http from "./http-common";
-
-export default function getAllEvents() {
-    return http.get(`/getAllEvents`);
-}
-
